Implement initprocess_X11 as Dummy
For every subsystem configuration
(at the current stage only X11/Opengl graphics subsystem)
the initialization process in initprocess_umgebung / freeprocess_umgebung
has been encoded differently. Textdb is used to make the different codings
not error-prone.

But now it suffices to have only one init and free encoding and for every
subsystem which is not included in the build process a dummy initprocess/free...
pair is defined in the header of the subsystem, e.g. in "x11.h":
   #if (KONFIG_GRAPHIK!=X11)
      #define initprocess_X11()  (0)
      #define freeprocess_X11()  (0)
   #endif