Add channel support to LOG output
A new enum type log_channel_e is introduced to support
different log channels.
At current point in time the values
    log_channel_ERR
and
    log_channel_TEST
are supported.

The log helper functions LOG_PRINTF, LOG_TEXTRES take an additional
argument to determine the logging channel.
It can be either set to ERR or TEST.