textdb: implemented (if (<COND>) (<select>))
The textdb precompiler if function is extended

Formerly only the function syntax (if (field != "") "value") was supported.
It writes the value "value" if the input field "field" is not empty.

The new syntax is now: (if (field != "") ("value" field \n "..." field2))
The output writtem is in case field is not empty is a linear combination of
strings and contents of database fields.