Implemented send_/wait_/trywait_rtsignal
Addes realtime signal support. Up to 16 (0..15)
signal numbers are accepted.

They can be send to the own process and any thread can
wait for them. If more than one thread is waiting for
the same signal number the receiver is chosen randomly.

Signals are stored in an internal queue which can overflow.
The size of the queue is dependent on the configuration of
the Linux kernel.