Implemented suspend_/resume_osthread
The thread impl. uses signals to wakeup another
thread which is suspended. This functionality allows
a thread to wait for a condition and it is met another
thread can wake up the waiting one.

The unittest now includes a test for inter thread singal handling
and another one for suspend_/resume_osthread.

The fields of struct osthread_t are reordered and renamed and
fields to support an upcoming implementation of waitlist_t ar added.