Implement waitlist_t in api/os/sync/
Use the waitlist to suspend a thread waiting for a certain condition.
Before a thread is woken up and removed from waitlist a command field
is set to the value given as parameter in the call to trywakeup_waitlist.

Other modules like the comming threadpool_t can use this command field
to transfer instructions to a thread before it is resumed.