Add unittest for test_errortimer_t
The unittest function is named unittest_test_errortimer.

test_errortimer_t is used to simulate an error in a function under test.
This is useful if a function calls other functions and you can not change
the called function or if you do not want to wrap every function call.