Implement arraysf_t, a (Patricia-)Trie index
arraysf_t stores (key, value) pairs.
The key is of type integer (size_t) and considered an index
into the array. The index is considered to be sparse.
Between the highest and lowest index number can be holes.
Unassigned indizes do not consume memory as is the case
for array allocated in a linear fashion.

It is implemented as a Patricia-trie with 4way branches.
Every internal trie node contains the bit position (from highest
to lowest) of the next 2 bits which are of interest.
According to their value the pointer to the next node is
chosen from an array of 4 pointers until a leaf node is found
where the value is stored.

From the root node to the leaf node only such bit positions are compared if there exist
at least two stored indizes which have different values at this position.
Therefore leaf nodes (type <arraysf_node_t>) has a depth of less than 16
(index_bitsize/2) on a 32 bit machine or 32 on a 64 bit machine.
The mean depth is log2(number of stored indizes)/2.

If you choose arraysf_MSBPOSROOT as the type of array the maximum depth for
a leaf node is log2(index value)/2 (depth<=4 for index values < 256).