Implement arraystf_t, a Trie index for strings
Trie implementation to support arrays index with string keys.

This type of array supports strings as indizes with arbitrary length.
Once an object is assigned a memory location the address is never changed (fixed location).
Every internal trie node contains a memory offset and a bit position (from highest to lowest).
If a key is searched for and an internal node is encountered the byte value at the memory offset in the key is taken and the value of the two bits are extracted.

The resuslting number (0..3) is taken as index into the child array of the internal
node <arraystf_mwaybranch_t>.

From the root node to the leaf node only such offset/bit combinations are compared if there exist at least two stored keys which have different values at this position.
Therefore leaf nodes (type <arraystf_node_t>) has a depth of less than (key length in bits)/2.
The mean depth is log2(number of stored keys)/2.

Special Encoding:
Cause keys of different length make problems in a patricia trie implementation if a smaller key is a prefix of a longer one.

A C-string has a null byte at its end so no stored string can be a prefix of another longer one.
To allows keys of any binary content an encoding of is chosen so that every key has the same length.
The encoding is as follows (keylength: key->size; keycontent: key->addr[0 .. key->size-1])

 - Every key is of length ~(size_t)0 (UINT32_MAX or UINT64_MAX)
 - The byte values at offset 0 .. key->size-1 are taken from key->addr
 - The byte values key->size .. UINT_MAX-1 are always 0
 - The value at offset is key->size and has 8*sizeof(size_t) bits instead of only 8