Implement transient memory manager
The memory manager is implemented as a new service in threadcontext_t.
Its initthread_mmtransient function is called every time a new thread
is created.

The current implementation of mm_transient_t is only a wrapper to the
standard library functions realloc and free.

In a later stage optimized memory handling (pooling) will be considered.