Move main/*test*_main.c into main/test/ subdir
The directory C-kern/main/ contains all main files.
In the directory C-kern/main/tools/ are all main files
for the tool projects like genmake and text resource compiler.

In the directory C-kern/main/test/ are now all main functions
used as startup functions for test projects.

There is a main module for running the unittest of every
module and a main module for a simple test childprocess
which is used in the unittest of the process module.