Allow mmtest_t to simulate errors.
* The function setresizeerr_mmtest and setfreeerr_mmtest
takes as parameter a pointer to a test_errortimer_t.

If a timer object is set the functions mresize_mmtest or mfree_mmtest
return an error code if the set timer object fires (the error code
is stored in the timer object).

The timer reference is cleared (set to 0) after it has fired.

* Rename switchteston_mmtest into switchon_mmtest;
  switchtestoff_mmtest is also renamed into switchoff_mmtest.