Add coding-conventions/error-handling.txt
The new behaviour of error-handling in functions is allowed.
Functions now can set the values of out parameters to a FREEABLE
state. Parameters which carry any additional return values can be
set to cleared state (0 in case of numbers).

This helps to simplify the implementation in some cases.
An advantage is that complex return values can now be constructed with help
of the return or out parameter.
In case of an error a simple clear_XXX or free_XXX resets all changes.
The caller should check the documentation if a function clears a
return parameter or free an out parameter in case of an error.