Implement dlist_t: double linked list
The list module is contained in the files
    C-kern/api/ds/inmem/node/dlist_node.h
    C-kern/api/ds/inmem/dlist.h
    C-kern/ds/inmem/dlist.c

This implementation also changes the way nodes are managed.
The nodes stored in dlist are of type dlist_node_t. The next and
and prev pointers point to dlist_node_t and not to some generic
object type. The generic implementation changes the pointers to
the correct object type with help of small wrapper functions.

A new typeadapt_t is introduced which combines the needed
lifetime service into a single object and which contains also
the offset from dlist_node_t to the container object.

It will replace typeadapter_t (the old implementation) step by step.