Make unittests agnostic to NR of open filedescr.
Some unittests write the nr of the open fd
into the error log. This change opens files
until 8 file descriptors are in use.

So the logged number is always 8 or higher.
This change was necessary to support X11 window tests.
The used libraries open 2 hidden file descriptors
unknown to the test program.