Change key_t in typeadapt_keycomparator_EMBED
The type of the parameter key_t in typeadapt_keycomparator_EMBED is changed.
The key_t type in other macros like typeadapt_keycomparator_DECLARE
and typeadapt_EMBED which uses typeadapt_keycomparator_EMBED are
therefore also changed.

The parameter key_t was used as (const key_t *) in the macro.
Now the macro uses (key_t) so all users of this macro has to
change the parameter key_t from void to void* or type_t to const type_t*.