Rename "_iot" type suffixes into "_t"
The interfaceable object type log_iot which is a structure
containing a pointer to an object and a pointer to the interface
log_it is renamed into log_t.

Also type mm_iot which contains a pointer to a memory manager object and a
pointer to its implementation of interface mm_it is renamed into mm_t.

Also type iocallback_iot which contains a pointer to an iohandler_t object
and a pointer to its implementation of iocallback_f function is renamed
into iocallback_t. Also a unittest for iocallback_t is added.

Type objectcache_iot which contains a pointer to an object of type objectcac
into objectcache_t. The default implementation of the old objectcache_iot
is renamed from objectcache_t into objectcache_impl_t.