Implement Extendible Hashing
The hash table has always a size which is a power of two.
It grows automatically (doubling its size) whenever the
number of entries in a certain slot (bucket) is exceeded.
To reduce the time needed for doubling the hash table
the newly created table entries share its content with
the old entries.

 ---------------
 | index[0b00] | --> (tree)    <---
 ---------------                  |
 | index[0b01] | --> (tree)    <--|---   (level 1 / 2 == pow(2, 1))
 ---------------                  |  |
 | index[0b10] | --> (shared)  ---|  |
 ---------------                     |
 | index[0b11] | --> (shared)  -------   (level 2 / 4 == pow(2, 2))
 ---------------