Add support for string table in textrescompiler
The textresource_compiler now supports the generation
of a string table which lists all text resources.

Conditional text and parameter values are not supported
in case of string table generation.

The version of the new resource format is "3".

A test table "C-kern/resource/errtab.text" is added to the resource
section which describes the new format.