Adapt tool generrtab to generate textresource
Generrtab was changed to write its output in a form
which is is understood by the textresource_compiler.

It writes its output also in two languages.
The language defined in the user environment (which
is German in my case) and the default language of the
C environment (which is English).