Implement blockarray_t
A blockarray is an array where not all
elements are stored in a single block of memory.
The type blockarray_t uses a configurable block size
and organizes the blocks in a B-tree like structure
where blocks containing data elements are the bottom level
(depth 0) and blocks contains ptrs to other blocks are mid
or top level (depth > 0 and root).

Added function releasecached to pagecache_t
and implemented releasecached_pagecacheimpl.
The function returns all cached unused memory blocks
back to the operating system. This is needed
to make same_resourceusage working in unittest_ds_inmem_blockarray.