Introduce iobj_t
This always available type (included in konfig.h)
offers macros to construct interfaceable objects.

The standard structure of an interfaceable object
are two pointers. The first points to the object data.
The second points to the function table implementing
the interface of the object.

   struct objtype_t {
      struct objtype_t  * object ;
      struct objtype_it * iimpl ;
   } ;

This definition allows any object without introducing
any additional "vtable" pointer to be used through an
interface. This definition is not useful for high
volume interfaceable objects but it is for objects
used as a single instance per thread.

------

Also refactored part of konfig.h into "C-kern/api/context/stdtypes.h".