Implement rwlock_t
The type rwlock_t offers a simple many reader
and single writer lock.

The implementation uses two mutex_t namely entrylock and exitlock.
The entrylock must be acquired by every reader or writer
before it is allowed to enter the protected region.
If a writer acquires the entrylock it holds it until it leaves
the protected region. It acquires also the exitlock sets
rwlock_t.writer to itself and checks rwlock_t.nrofreader.
It releases exitlock and suspends itself if nrofreader was != 0.
If a reader enters the protected region it acquires entrylock
adds 1 to nrofreader and releases it immediately.
Thus entrylock ensures exclusive access to the protected region.

If a reader leaves the protected region it acquires exitlock,
decreases nrofreader and resumes a writer if nrofreader reached 0.
It releases exitlock. If a writer has been resumed it acquires exitlock
checks that there is no reader and then it enters the protected region
else it suspends itself again. This is necessary cause suspend and resume
are implemented with signals and it is possible that a spurious signal
is received from outside the process and therefore the writer has to check
the condition again.