Implement Thread Local Storage
The module thread_tls.c is the result of the refactoring of
stackframe implementation from module thread.

It allocates memory for all thread local variables but does not
use it. The replacement of all variables declared with __thread
as part of the thread local storage will be done in a future step.