Implement startup_platform
startup_platform does a platform specific startup initialization
and calls the implementation of the main thread.
It is implemented in file C-kern/platform/Linux/startup.c.

THe main thread calls init_maincontext before calling
any other functions of the C-kern system and it calls
free_maincontext before exit - as usual.

startup_platform allocates a stack and runs the main thread
of the process on this stack which is compatible to thread_tls_t.

This allows thread_tls_t to locate the thread local variables
relative to the address of the current stack.

In a later step thread local storage supported by compiler
(keyword __thread) will be replaced with help of thread_tls_t.