Access processcontext_t without global variable
The global processcontext_t was accessed with g_maincontext.pcontext.
This was changed and now it is accessed with tcontext_maincontext()->pcontex

tcontext_maincontext() returns the current threadcontext_t which was
extended with pointer pcontext which in turn points to the shared
processcontext.

This allows modules to access the shared processcontext_t without
use of any global variable !