Rename interfacethread_TYPE into interface_TYPE
For example interfacethread_objectcacheimpl was renamed
into interface_objectcacheimpl. The interface functions
return the interface of a type which is initialized for
use in the thread global threadcontext_t.

Also changed the initializaion in init_threadcontext to
check for interface_TYPE return values of 0.
If a 0 value is returned the static object is kept.
This corrected an error where logging did no more work in case
of single threaded applications.
logwriter_t is only used in case KONFIG_SUBSYS
contains KONFIG_thread (threads are used). If it is not
used (interface_logwriter returns 0) the static logging service
logmain_t is kept.