utf8scanner_t handles chars split acrosss buffers
The interface of utf8scanner_t was changed.
Instead of starting and ending a token explicitly
scanning the buffer automatically starts a token and
calling scannedtoken_utf8scanner ends the token.

Added new functions to scan for utf8 characters.
Previously only bytes could be read.
If a multibyte sequence is split over two buffers
it is handled correctly.