Introduce ramsize_t
Implement functions
 - sizeavailableram_vm()
 - and sizephysram_vm()
which determine the size of the available physical memory in bytes
and the total size of physical memory. The returned numbers has
type ramsize_t (64bit) cause the size of ram could exceed 4GB
(32 bits) on a 32 bit system.

Unittest was added to check that sizeof(ramsize_t) >= sizeof(size_t)
in case of a 128 bit machine !