Add type uintfct_t
This integer type is big enough to hold a pointer to a function.
POSIX interfaces assume sizeof(void*) == sizeof(void(*)(void))
but C allows pointers to functions of a different size than pointer
to objects.

The default definition also assumes sizeof(void*)==sizeof(void(*)())
but an added unittest checks that sizeof(uintfct_t) is big enough.

In case of some special platform a system specific sys_uintfct_t
has to be added and uintfct_t has to be defined as an alias.