Change log_it interface
The funtion printf from log_it has a new parameter which is a pointer
to log_header_t a newly introduced type.
If this parameter is != 0 a log header which contains timestamp and
error location (function, file, linenr) and error nr is written.
The header replaces log_flags_TIMESTAMP which was removed.

The functions flushbuffer, clearbuffer, and getbuffer are extended with
an additional channel parameter.