Implement filepath_static_t
Implemented filepath_static_t which describes the path of a working
directory and the filename or filepath relative to the working
directory.

If the filename is an absolute path the path of the working directory
is ignored and only filename is stored.

This type is used to log the path of failed "file open" operations.