Add log_state_e to log interface
The log_it interface was extended with two function:

 *  log_state_e getstate(const void * log, log_channel_e channel) ;

 *  void setstate(void * log, log_channel_e channel, log_state_e logstate) ;

With these two functions every log channel can be configured to ignore
any write attempts, or buffer the log and write only if buffer full,
or to write every log entry immediately.

Added new flag log_flags_OPTIONALHEADER which writes the header only if this
is the first header written in the function. The log function remembers
the function name stored in the header. The flag log_flags_OPTIONALHEADER
ensures that the header is only written if the function name is different.
The flag log_flags_END clears the remembered function name in the log object.
Therefore every function must contain a log output at its end with
log_flags_END set. To ensure this TRACEABORT and TRACEABORTFREE which
logs the failure of a function and which are logged last set this flag.