Implement daemonize_process
The function daemonize_process replaces initdaemon_process.
The difference is that daemonize_process makes the calling
process a daemon process whereas initdaemon_process created
a new process running as a deamon.

This allows the runnig process to become a daemon without
creating another process. Internally fork() so
daemonize_process creates a child process and ends the caller.
But the it appears as if the caller is daemonized.