Implement memory manager MM_TEST macros
Use ALLOC_MM_TEST instead of ALLOC_MM to generate an error in case of
KONFIG_UNITTEST is defined.
Use RESIZE_MM_TEST instead of RESIZE_MM to generate an error in case of
KONFIG_UNITTEST is defined.
Use FREE_MM_TEST instead of FREE_MM to generate an error in case of
KONFIG_UNITTEST is defined.