Implement eglwindow_t
An eglwindow_t wraps a native window (X11 window)
into an EGL window surface.
This surface can be used as target for drawing
commands issued by OpenGL or OpenGL ES.

glxwindow_t is removed.
Instead EGL is the preferred way to initialize OpenGL.
If necessary GLX is reintroduced later but with
a new interface which integrates with surface_config_t
the same way EGL does.