Implement a simple OpenGL-ES/EGL Demo
This demo sets up an X11 window which
can be used with OpenGL-ES.
The setup between native windows and
OpenGL-ES is done with EGL.

The OpenGL-ES setup allows you to draw
in »Normalized Device Coordinates(NDC)«
(w coordinate is assumed to be 1),
ranging from -1 to 1 in every dimension.
The code draws two rectangles which overlap.

The demo code can be started with
> bin/demo 1
or
> bin/demo setup_opengles